Adicionado NotAllowedOperationException

+++ src/main/java/sgallalucas/fitlab/services/StudentService.java
@@ -3,8 +3,10 @@ package sgallalucas.fitlab.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import sgallalucas.fitlab.dtos.StudentDto;
+import sgallalucas.fitlab.exceptions.NotAllowedOperationException;
 import sgallalucas.fitlab.model.Student;
 import sgallalucas.fitlab.repositories.StudentRepository;
+import sgallalucas.fitlab.repositories.WorkoutRepository;
 import sgallalucas.fitlab.validators.StudentValidator;
 
 import java.util.List;
@@ -15,6 +17,7 @@ import java.util.UUID;
 public class StudentService {
 
     private final StudentRepository repository;
+    private final WorkoutRepository workoutRepository;
     private final StudentValidator validator;
 
     public Student save(Student student) {
@@ -32,11 +35,19 @@ public class StudentService {
     }
 
     public void update(Student student) {
+        validator.validation(student.getEmail());
         repository.save(student);
     }
 
     public void delete(Student student) {
+        if (hasWorkout(student)) {
+            throw new NotAllowedOperationException("It is not allowed delete a student who has workouts");
+        }
         repository.delete(student);
+    }
+
+    public boolean hasWorkout(Student student) {
+        return workoutRepository.existsByStudent(student);
     }
 
     public Student convertToEntity(StudentDto dto) {
@@ -46,7 +57,6 @@ public class StudentService {
         student.setBirthDate(dto.birthDate());
         student.setEmail((dto.email()));
         student.setGenre(dto.genre());
-
         return student;
     }
